# my-expense-tracker/old/users-api/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CreateUserRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      username:
        example: alice
        minLength: 3
        type: string
    required:
    - username
    type: object
  model.ExportSimpleRequest:
    properties:
      clear:
        type: boolean
      rows:
        items:
          $ref: '#/definitions/model.SimpleRow'
        minItems: 1
        type: array
      sheet_name:
        example: Report
        type: string
      spreadsheet_id:
        example: 1AbCDefGhijk
        type: string
    required:
    - rows
    - sheet_name
    - spreadsheet_id
    type: object
  model.ExportSimpleResponse:
    properties:
      rows:
        example: 3
        type: integer
      spreadsheet_url:
        example: https://docs.google.com/spreadsheets/d/<ID>/edit
        type: string
      written_range:
        example: Report!A1:B4
        type: string
    type: object
  model.SimpleRow:
    properties:
      amount:
        example: 1250.5
        type: number
      category:
        example: Продукты
        type: string
    type: object
  model.UpdateUserRequest:
    properties:
      email:
        example: alice2@example.com
        type: string
      username:
        example: alice2
        type: string
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Пример API пользователей на Gin
  title: Users API
  version: "1.0"
paths:
  /demo/download:
    get:
      description: Считывает столбцы A/B из таблицы, указанной в переменных окружения
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/model.SimpleRow'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download (без параметров)
      tags:
      - demo
  /demo/export:
    post:
      consumes:
      - application/json
      description: Пишет три строки с категориями и суммами в указанный лист
      parameters:
      - description: Настройки листа
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ExportSimpleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ExportSimpleResponse'
      summary: Заполнить таблицу демо-данными
      tags:
      - demo
  /demo/import:
    get:
      description: Возвращает все заполненные строки из колонок A и B
      parameters:
      - description: ID таблицы
        in: query
        name: spreadsheet_id
        required: true
        type: string
      - default: Report
        description: Название листа
        in: query
        name: sheet_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/model.SimpleRow'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Прочитать значения из двух колонок
      tags:
      - demo
  /users:
    get:
      description: Возвращает список пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UserResponse'
            type: array
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: ID пользователя
        example: 11111111-1111-1111-1111-111111111111
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает пользователя по ID
      parameters:
      - description: ID пользователя
        example: 11111111-1111-1111-1111-111111111111
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Частично обновляет данные пользователя
      parameters:
      - description: ID пользователя
        example: 11111111-1111-1111-1111-111111111111
        in: path
        name: id
        required: true
        type: string
      - description: Изменяемые поля
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить пользователя
      tags:
      - users
schemes:
- http
swagger: "2.0"